django_diplom/requirements.txt
asgiref==3.4.1
certifi==2021.10.8
charset-normalizer==2.0.7
defusedxml==0.7.1
Django==3.2.7
django-colorfield==0.4.5
django-crispy-forms==1.13.0
django-widget-tweaks==1.4.8
idna==3.3
javascript==1!0.2.12
oauthlib==3.1.1
Pillow==8.3.2
PyJWT==2.3.0
python-social-auth==0.2.12
python3-openid==3.2.0
pytz==2021.1
requests==2.26.0
requests-oauthlib==1.3.0
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.7
